+++ packages/localizations/src/esUs.ts
@@ -108,6 +108,7 @@ export const esUs = {
       paze: 'Paze',
       offline: 'Pagos offline',
       ach: 'Cuenta Bancaria',
+      mercadopago: 'Mercado Pago',
     },
     descriptions: {
       creditCard: '',
@@ -117,6 +118,8 @@ export const esUs = {
       paze: '',
       offline: '',
       ach: '',
+      mercadopago:
+        'Usa el formulario de MercadoPago a continuación para completar tu compra de forma segura.',
     },
     noMethodsAvailable: 'No hay métodos de pago disponibles',
     cardNumber: 'Número de tarjeta',