+++ packages/localizations/src/zhCn.ts
@@ -103,6 +103,7 @@ export const zhCn = {
       paze: 'Paze',
       offline: '线下付款',
       ach: '银行账户',
+      mercadopago: 'Mercado Pago',
     },
     descriptions: {
       creditCard: '',
@@ -112,6 +113,7 @@ export const zhCn = {
       paze: '',
       offline: '',
       ach: '',
+      mercadopago: '请使用下方的 MercadoPago 表单安全完成购买。',
     },
     noMethodsAvailable: '暂无可用的付款方式',
     cardNumber: '卡号',