+++ packages/react/src/components/checkout/payment/lazy-payment-loader.tsx
@@ -88,6 +88,13 @@ const LazyComponents = {
     )
   ),
 
+  MercadoPagoCheckoutButton: lazy(() =>
+    import(
+      '@/components/checkout/payment/checkout-buttons/mercadopago/mercadopago'
+    ).then(module => ({
+      default: module.MercadoPagoCheckoutButton,
+    }))
+  ),
   // Express Buttons
   ExpressCheckoutButton: lazy(() =>
     import(
@@ -202,6 +209,11 @@ type PaymentComponentRegistry = {
       button: PaymentComponentKey;
     };
   };
+  [PaymentMethodType.MERCADOPAGO]?: {
+    [PaymentProvider.MERCADOPAGO]: {
+      button: PaymentComponentKey;
+    };
+  };
 };
 
 export const lazyPaymentComponentRegistry: PaymentComponentRegistry = {
@@ -249,6 +261,11 @@ export const lazyPaymentComponentRegistry: PaymentComponentRegistry = {
   [PaymentMethodType.PAZE]: {
     [PaymentProvider.GODADDY]: {
       button: 'PazeCheckoutButton',
+    },
+  },
+  [PaymentMethodType.MERCADOPAGO]: {
+    [PaymentProvider.MERCADOPAGO]: {
+      button: 'MercadoPagoCheckoutButton',
     },
   },
   [PaymentMethodType.ACH]: {