+++ packages/react/src/types.ts
@@ -43,6 +43,7 @@ export const PaymentProvider = {
   PAYPAL: 'paypal',
   PAZE: 'paze',
   OFFLINE: 'offline',
+  MERCADOPAGO: 'mercadopago',
 } as const;
 
 export const CheckoutType = {
@@ -63,6 +64,7 @@ export const PaymentMethodType = {
   GOOGLE_PAY: 'googlePay',
   OFFLINE: 'offline',
   PAZE: 'paze',
+  MERCADOPAGO: 'mercadopago',
 } as const;
 
 // Union of all payment method keys